fix some bug in the tujs

+++ components/components-jdbc/components-jdbc-runtime/src/main/java/org/talend/components/jdbc/runtime/type/RowWriter.java
@@ -250,7 +250,7 @@ public class RowWriter {
             if (inputValue == null) {
                 statement.setNull(statementIndex, java.sql.Types.INTEGER);
             } else {
-                statement.setShort(statementIndex, (short) inputValue);
+                statement.setShort(statementIndex, ((Integer) inputValue).shortValue());
             }
         }
 
@@ -267,7 +267,9 @@ public class RowWriter {
             if (inputValue == null) {
                 statement.setNull(statementIndex, java.sql.Types.INTEGER);
             } else {
-                statement.setByte(statementIndex, (byte) inputValue);
+                //please see org.talend.codegen.enforcer.IncomingSchemaEnforcer, it will convert byte(Byte) to int(Integer), not know why, so change here
+                //statement.setByte(statementIndex, (byte) inputValue);
+                statement.setByte(statementIndex, ((Integer) inputValue).byteValue());
             }
         }
